Moved Committee into its own file
